+++ apps/api/src/app/info/info.module.ts
@@ -1,10 +1,9 @@
 import { RedisCacheModule } from '@ghostfolio/api/app/redis-cache/redis-cache.module';
-import { ConfigurationService } from '@ghostfolio/api/services/configuration.service';
+import { ConfigurationModule } from '@ghostfolio/api/services/configuration.module';
 import { DataGatheringModule } from '@ghostfolio/api/services/data-gathering.module';
-import { DataGatheringService } from '@ghostfolio/api/services/data-gathering.service';
 import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
 import { ExchangeRateDataModule } from '@ghostfolio/api/services/exchange-rate-data.module';
-import { PrismaService } from '@ghostfolio/api/services/prisma.service';
+import { PrismaModule } from '@ghostfolio/api/services/prisma.module';
 import { PropertyModule } from '@ghostfolio/api/services/property/property.module';
 import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile.module';
 import { Module } from '@nestjs/common';
@@ -14,7 +13,9 @@ import { InfoController } from './info.controller';
 import { InfoService } from './info.service';
 
 @Module({
+  controllers: [InfoController],
   imports: [
+    ConfigurationModule,
     DataGatheringModule,
     DataProviderModule,
     ExchangeRateDataModule,
@@ -22,16 +23,11 @@ import { InfoService } from './info.service';
       secret: process.env.JWT_SECRET_KEY,
       signOptions: { expiresIn: '30 days' }
     }),
+    PrismaModule,
     PropertyModule,
     RedisCacheModule,
     SymbolProfileModule
   ],
-  controllers: [InfoController],
-  providers: [
-    ConfigurationService,
-    DataGatheringService,
-    InfoService,
-    PrismaService
-  ]
+  providers: [InfoService]
 })
 export class InfoModule {}