+++ apps/api/src/app/order/order.module.ts
@@ -1,5 +1,5 @@
 import { AccountService } from '@ghostfolio/api/app/account/account.service';
-import { CacheService } from '@ghostfolio/api/app/cache/cache.service';
+import { CacheModule } from '@ghostfolio/api/app/cache/cache.module';
 import { RedisCacheModule } from '@ghostfolio/api/app/redis-cache/redis-cache.module';
 import { UserModule } from '@ghostfolio/api/app/user/user.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration.module';
@@ -15,7 +15,10 @@ import { OrderController } from './order.controller';
 import { OrderService } from './order.service';
 
 @Module({
+  controllers: [OrderController],
+  exports: [OrderService],
   imports: [
+    CacheModule,
     ConfigurationModule,
     DataGatheringModule,
     DataProviderModule,
@@ -26,8 +29,6 @@ import { OrderService } from './order.service';
     SymbolProfileModule,
     UserModule
   ],
-  controllers: [OrderController],
-  providers: [AccountService, CacheService, OrderService],
-  exports: [OrderService]
+  providers: [AccountService, OrderService]
 })
 export class OrderModule {}