+++ apps/api/src/app/portfolio/portfolio.module.ts
@@ -20,6 +20,7 @@ import { PortfolioServiceNew } from './portfolio.service-new';
 import { RulesService } from './rules.service';
 
 @Module({
+  controllers: [PortfolioController],
   exports: [PortfolioServiceStrategy],
   imports: [
     AccessModule,
@@ -34,7 +35,6 @@ import { RulesService } from './rules.service';
     SymbolProfileModule,
     UserModule
   ],
-  controllers: [PortfolioController],
   providers: [
     AccountService,
     CurrentRateService,