+++ apps/api/src/app/subscription/subscription.module.ts
@@ -1,5 +1,5 @@
-import { ConfigurationService } from '@ghostfolio/api/services/configuration.service';
-import { PrismaService } from '@ghostfolio/api/services/prisma.service';
+import { ConfigurationModule } from '@ghostfolio/api/services/configuration.module';
+import { PrismaModule } from '@ghostfolio/api/services/prisma.module';
 import { PropertyModule } from '@ghostfolio/api/services/property/property.module';
 import { Module } from '@nestjs/common';
 
@@ -7,9 +7,9 @@ import { SubscriptionController } from './subscription.controller';
 import { SubscriptionService } from './subscription.service';
 
 @Module({
-  imports: [PropertyModule],
   controllers: [SubscriptionController],
-  providers: [ConfigurationService, PrismaService, SubscriptionService],
-  exports: [SubscriptionService]
+  exports: [SubscriptionService],
+  imports: [ConfigurationModule, PrismaModule, PropertyModule],
+  providers: [SubscriptionService]
 })
 export class SubscriptionModule {}