+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved the countries and sectors charts in the position detail dialog
 - Restructured the server modules
 
+### Fixed
+
+- Fixed the allocations by account for non-unique account names
+
 ## 1.116.0 - 16.02.2022
 
 ### Added